Adding sequence types.

+++ library/thrust_wrappers/index_sequence.h
@@ -2,9 +2,9 @@
 #include "iterator_sequence.h"
 #include <thrust/iterator/counting_iterator.h>
 
-struct index_sequence : public iterator_sequence<thrust::counting_iterator<int> >
+struct index_sequence : public iterator_sequence<thrust::counting_iterator<long> >
 {
-  __host__ __device__ index_sequence(int length) :
-  iterator_sequence<thrust::counting_iterator<int> >(thrust::counting_iterator<int>(0), length) {}
+  __host__ __device__ index_sequence(long length) :
+  iterator_sequence<thrust::counting_iterator<long> >(thrust::counting_iterator<int>(0), length) {}
 };
